inline_scripts in pop_jobs and pop_categories
Adding support for inline_scripts to the pop_jobs and pop_categories folders in the common directory. 
The vanilla files make use of them in job = {} and resources = {} so I added support for them in those fields. 
Inline scripts require relative file path in common/inline_scripts/ folder. It is also possible to create sub-folders in the inline_scripts folder to organize various scripts.