3.10 changed how the modifiers worked to get scopes correct
The old way for the triggered modifiers was not working with the scopes and there were a number of new triggered modifiers.   The way it is now takes more typing but seems to work and is easier to understand for a non expert.  Sorry but my head hurts from trying to figure out how the old way worked.  The old way might look better for a programmer but the people who do the updates need something they can follow and update.