Update projectiles.cwt
Fix warnings for color and width etc.   Very frustrating to get the warnings to go away.   Note the last color number can be above 1 as paradox uses 5.  I can not find a way to specify it is only the last color number that can be above 1 so I had to allow any of them to be above 1.